+++ src/sdk/PnP.Core.Test/SharePoint/ListItemTests.cs
@@ -1,14 +1,13 @@
-﻿using AngleSharp.Dom;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using PnP.Core.Model;
 using PnP.Core.Model.Security;
 using PnP.Core.Model.SharePoint;
+using PnP.Core.QueryModel;
 using PnP.Core.Test.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using PnP.Core.QueryModel;
-using PnP.Core.Model;
 
 namespace PnP.Core.Test.SharePoint
 {