+++ src/sdk/PnP.Core/Model/SharePoint/Core/Internal/ListItem.cs
@@ -22,6 +22,7 @@ namespace PnP.Core.Model.SharePoint
     /// </summary>
     [SharePointType("SP.ListItem", Target = typeof(List), Uri = "_api/web/lists/getbyid(guid'{Parent.Id}')/items({Id})", LinqGet = "_api/web/lists(guid'{Parent.Id}')/items")]
     [SharePointType("SP.ListItem", Target = typeof(File), Uri = "_api/web/getFileById('{Parent.Id}')/listitemallfields")]
+    [SharePointType("SP.ListItem", Target = typeof(Folder), Uri = "_api/Web/getFolderById('{Parent.Id}')/listitemallfields")]
     //[GraphType(OverflowProperty = "fields")]
     internal partial class ListItem : ExpandoBaseDataModel<IListItem>, IListItem
     {
@@ -196,6 +197,24 @@ namespace PnP.Core.Model.SharePoint
 
         #region Methods
 
+        #region Common
+
+        private string GetItemUri()
+        {
+            if (Parent is Folder)
+            {
+                return "_api/Web/getFolderById('{Parent.Id}')/listitemallfields";
+            }
+            else if (Parent is File)
+            {
+                return "_api/web/getFileById('{Parent.Id}')/listitemallfields";
+            }
+
+            return "_api/Web/lists/getbyid(guid'{Parent.Id}')/items({Id})";
+        }
+
+        #endregion
+
         #region Folder
         public async Task<bool> IsFolderAsync()
         {
@@ -229,7 +248,7 @@ namespace PnP.Core.Model.SharePoint
 
         private async Task LoadKeyListItemProperties()
         {
-            ApiCall apiCall = new ApiCall($"_api/web/lists/getbyid(guid'{{Parent.Id}}')/items({Id})?$select=ContentTypeId,FileDirRef", ApiType.SPORest);
+            ApiCall apiCall = new ApiCall($"{GetItemUri()}?$select=ContentTypeId,FileDirRef", ApiType.SPORest);
             await RequestAsync(apiCall, HttpMethod.Get).ConfigureAwait(false);
         }
         #endregion
@@ -578,7 +597,7 @@ namespace PnP.Core.Model.SharePoint
 
         public async Task<bool> AreCommentsDisabledAsync()
         {
-            var apiCall = new ApiCall("_api/web/lists/getbyid(guid'{Parent.Id}')/items({Id})/CommentsDisabled", ApiType.SPORest);
+            var apiCall = new ApiCall($"{GetItemUri()}/CommentsDisabled", ApiType.SPORest);
 
             var response = await RawRequestAsync(apiCall, HttpMethod.Post).ConfigureAwait(false);
 
@@ -610,7 +629,7 @@ namespace PnP.Core.Model.SharePoint
 
             string body = JsonSerializer.Serialize(parameters);
 
-            var apiCall = new ApiCall("_api/web/lists/getbyid(guid'{Parent.Id}')/items({Id})/SetCommentsDisabled", ApiType.SPORest, body);
+            var apiCall = new ApiCall($"{GetItemUri()}/SetCommentsDisabled", ApiType.SPORest, body);
 
             await RawRequestAsync(apiCall, HttpMethod.Post).ConfigureAwait(false);
         }
@@ -651,7 +670,7 @@ namespace PnP.Core.Model.SharePoint
             await RawRequestBatchAsync(batch, apiCall, HttpMethod.Post).ConfigureAwait(false);
         }
 
-        private static ApiCall SetComplianceTagApiCall(string complianceTag, bool isTagPolicyHold, bool isTagPolicyRecord, bool isEventBasedTag, bool isTagSuperLock)
+        private ApiCall SetComplianceTagApiCall(string complianceTag, bool isTagPolicyHold, bool isTagPolicyRecord, bool isEventBasedTag, bool isTagSuperLock)
         {
             var parameters = new
             {
@@ -662,7 +681,7 @@ namespace PnP.Core.Model.SharePoint
                 isTagSuperLock
             };
             string body = JsonSerializer.Serialize(parameters);
-            var apiCall = new ApiCall("_api/web/lists/getbyid(guid'{Parent.Id}')/items({Id})/SetComplianceTag", ApiType.SPORest, body);
+            var apiCall = new ApiCall($"{GetItemUri()}/SetComplianceTag", ApiType.SPORest, body);
             return apiCall;
         }
         #endregion
@@ -730,9 +749,9 @@ namespace PnP.Core.Model.SharePoint
             return new BatchSingleResult<BatchResultValue<Guid>>(batch, batchRequest.Id, apiCall.RawSingleResult as BatchResultValue<Guid>);
         }
 
-        private static ApiCall BuildRecycleApiCall()
+        private ApiCall BuildRecycleApiCall()
         {
-            return new ApiCall("_api/Web/Lists(guid'{Parent.Id}')/items({Id})/recycle", ApiType.SPORest)
+            return new ApiCall($"{GetItemUri()}/recycle", ApiType.SPORest)
             {
                 RemoveFromModel = true
             };
@@ -906,8 +925,34 @@ namespace PnP.Core.Model.SharePoint
 
         public async Task BreakRoleInheritanceAsync(bool copyRoleAssignments, bool clearSubscopes)
         {
-            var apiCall = new ApiCall("_api/Web/Lists(guid'{Parent.Id}')/items({Id})/BreakRoleInheritance(copyRoleAssignments=" + copyRoleAssignments.ToString().ToLower() + ",clearSubscopes=" + clearSubscopes.ToString().ToLower() + ")", ApiType.SPORest);
+            ApiCall apiCall = BuildBreakRoleInheritanceApiCall(copyRoleAssignments, clearSubscopes);
             await RawRequestAsync(apiCall, HttpMethod.Post).ConfigureAwait(false);
+        }
+
+        public void BreakRoleInheritanceBatch(Batch batch, bool copyRoleAssignments, bool clearSubscopes)
+        {
+            BreakRoleInheritanceBatchAsync(batch, copyRoleAssignments, clearSubscopes).GetAwaiter().GetResult();
+        }
+
+        public async Task BreakRoleInheritanceBatchAsync(Batch batch, bool copyRoleAssignments, bool clearSubscopes)
+        {
+            ApiCall apiCall = BuildBreakRoleInheritanceApiCall(copyRoleAssignments, clearSubscopes);
+            await RawRequestBatchAsync(batch, apiCall, HttpMethod.Post).ConfigureAwait(false);
+        }
+
+        public void BreakRoleInheritanceBatch(bool copyRoleAssignments, bool clearSubscopes)
+        {
+            BreakRoleInheritanceBatchAsync(copyRoleAssignments, clearSubscopes).GetAwaiter().GetResult();
+        }
+
+        public async Task BreakRoleInheritanceBatchAsync(bool copyRoleAssignments, bool clearSubscopes)
+        {
+            await BreakRoleInheritanceBatchAsync(PnPContext.CurrentBatch, copyRoleAssignments, clearSubscopes).ConfigureAwait(false);
+        }
+
+        private ApiCall BuildBreakRoleInheritanceApiCall(bool copyRoleAssignments, bool clearSubscopes)
+        {
+            return new ApiCall($"{GetItemUri()}/BreakRoleInheritance(copyRoleAssignments=" + copyRoleAssignments.ToString().ToLower() + ",clearSubscopes=" + clearSubscopes.ToString().ToLower() + ")", ApiType.SPORest);
         }
 
         public void ResetRoleInheritance()
@@ -917,8 +962,34 @@ namespace PnP.Core.Model.SharePoint
 
         public async Task ResetRoleInheritanceAsync()
         {
-            var apiCall = new ApiCall("_api/Web/Lists(guid'{Parent.Id}')/items({Id})/ResetRoleInheritance", ApiType.SPORest);
+            ApiCall apiCall = BuildResetRoleInheritanceApiCall();
             await RawRequestAsync(apiCall, HttpMethod.Post).ConfigureAwait(false);
+        }
+
+        public void ResetRoleInheritanceBatch(Batch batch)
+        {
+            ResetRoleInheritanceBatchAsync(batch).GetAwaiter().GetResult();
+        }
+
+        public async Task ResetRoleInheritanceBatchAsync(Batch batch)
+        {
+            ApiCall apiCall = BuildResetRoleInheritanceApiCall();
+            await RawRequestBatchAsync(batch, apiCall, HttpMethod.Post).ConfigureAwait(false);
+        }
+
+        public void ResetRoleInheritanceBatch()
+        {
+            ResetRoleInheritanceBatchAsync().GetAwaiter().GetResult();
+        }
+
+        public async Task ResetRoleInheritanceBatchAsync()
+        {
+            await ResetRoleInheritanceBatchAsync(PnPContext.CurrentBatch).ConfigureAwait(false);
+        }
+
+        private ApiCall BuildResetRoleInheritanceApiCall()
+        {
+            return new ApiCall($"{GetItemUri()}/ResetRoleInheritance", ApiType.SPORest);
         }
 
         public IRoleDefinitionCollection GetRoleDefinitions(int principalId)
@@ -950,7 +1021,7 @@ namespace PnP.Core.Model.SharePoint
                 var roleDefinition = await PnPContext.Web.RoleDefinitions.FirstOrDefaultAsync(d => d.Name == name).ConfigureAwait(false);
                 if (roleDefinition != null)
                 {
-                    var apiCall = new ApiCall($"_api/web/lists(guid'{parentList.Id}')/Items({Id})/roleassignments/addroleassignment(principalid={principalId},roledefid={roleDefinition.Id})", ApiType.SPORest);
+                    var apiCall = new ApiCall($"{GetItemUri()}/roleassignments/addroleassignment(principalid={principalId},roledefid={roleDefinition.Id})", ApiType.SPORest);
                     await RawRequestAsync(apiCall, HttpMethod.Post).ConfigureAwait(false);
                     return true;
                 }
@@ -978,7 +1049,7 @@ namespace PnP.Core.Model.SharePoint
                 var roleDefinition = roleDefinitions.AsRequested().FirstOrDefault(r => r.Name == name);
                 if (roleDefinition != null)
                 {
-                    var apiCall = new ApiCall($"_api/web/lists(guid'{parentList.Id}')/Items({Id})/roleassignments/removeroleassignment(principalid={principalId},roledefid={roleDefinition.Id})", ApiType.SPORest);
+                    var apiCall = new ApiCall($"{GetItemUri()}/roleassignments/removeroleassignment(principalid={principalId},roledefid={roleDefinition.Id})", ApiType.SPORest);
                     await RawRequestAsync(apiCall, HttpMethod.Post).ConfigureAwait(false);
                     return true;
                 }